refactor(866): move Tickets, Budget and Gate jobs to their sections (G5 lane B2)
Four jobs leave `Humans.Infrastructure/Jobs`, following lane 5b-1's pattern: the job
lands in its owning section's `Contracts/` folder, `public`, taking the folder's
namespace. No registration site moved, no job id changed.

Ownership was re-measured from each body rather than the filename, and one of the four
does not belong where the lane brief guessed:

- `TicketSyncJob` -> `Humans.Tickets/Contracts/`. Drives `ITicketSync`; the manual
  trigger is Tickets' own `TicketController`.
- `TicketingBudgetSyncJob` -> `Humans.Budget/Contracts/`, **not Tickets**. Both
  collaborators are Budget's (`ITicketingBudgetService`, `IBudgetServiceRead`), the rows
  it writes are budget line items, and its job id is already `budget-ticketing-sync`. The
  4a placement inventory says Budget too; only the lane brief's paraphrase said Tickets.
- `GateRetentionJob` -> `Humans.Gate/Contracts/`. Purges `gate_scan_events` through
  `IGateScanRetention`.
- `GateVendorCheckInJob` -> `Humans.Gate/Contracts/`. Gate, not Tickets: both enqueue
  sites are Gate controllers (`GateController`, `GateVendorBackfillAdminController`), it
  writes no table at all, and its one collaborator is reached through Tickets' public
  `ITicketVendorMirror` - a normal section-to-section call.

`public` under `Contracts/` is forced, not chosen: Shell names each concrete type at two
sites (`AddScoped<T>()` and the recurring roll-call or a controller's `Enqueue<T>`) and
HUM0034 makes any other public type in a section assembly a build error.

Reference changes are a net simplification of Base. Two `<ProjectReference>` leave
`Humans.Infrastructure` entirely - `Humans.Budget.Contracts` and `Humans.Gate.Contracts`
each had exactly one consumer here, and it was the job that left. One is added,
`Humans.Gate -> Humans.Tickets.Contracts`, replacing the edge Base used to carry for
`GateVendorCheckInJob`; the leaf references `Humans.Interfaces` only, so no cycle is
possible. One package reference added: `Hangfire.Core` on `Humans.Budget` for
`[DisableConcurrentExecution]` - package, not project, and it was already arriving
transitively.

`Humans.Gate.Contracts` now has no consumer outside `Humans.Gate` at all: it was carved
out for `GateRetentionJob` alone. Folding `IGateScanRetention` into the section's own
`Contracts/` folder and deleting the project is a live follow-up, recorded in the leaf's
own csproj and in Gate.md but not done here - retiring a project is Peter's call.
`ITicketingBudgetService` is section-internal in fact for the same reason and was left
public rather than shrunk, matching lane 5b-1's open question on interface surface.

Comment and doc drift the move caused is corrected alongside: three `Section.cs`
remarks, the four interface docs that claimed their driver "stays in
`Humans.Infrastructure/Jobs`", the three contracts-leaf csproj rationales that counted
the departing jobs among the Base consumers justifying the leaf, the Gate/Budget/Tickets
section and feature docs, and four `freshness:triggers` blocks that pinned the old job
paths. `background-jobs.md` gains the roll-call file as a trigger so its catalog stays
covered as the Jobs folder drains.

Refs nobodies-collective/Humans#866
