+++ docs/features/tickets/ticket-vendor-integration.md
@@ -6,7 +6,6 @@
   src/Humans.Web/Views/Welcome/**
   src/Sections/Humans.Campaigns/Domain/CampaignGrant.cs
   src/Sections/Humans.Tickets.Contracts/TicketConstants.cs
-  src/Humans.Infrastructure/Jobs/TicketSyncJob.cs
   src/Sections/Humans.Stripe/**
 -->
 <!-- freshness:flag-on-change
