+++ src/Humans.Web/Extensions/InfrastructureServiceCollectionExtensions.cs
@@ -7,6 +7,9 @@ using Humans.Application.Interfaces;
 using Humans.Application.Interfaces.Caching;
 using Humans.Application.Interfaces.GoogleIntegration;
 using Humans.Application.Interfaces.Repositories;
+using Humans.Budget.Contracts;
+using Humans.Gate.Contracts;
+using Humans.Tickets.Contracts;
 using Humans.Infrastructure.Caching;
 using Humans.Infrastructure.Configuration;
 using Humans.Infrastructure.Jobs;
