+++ src/Sections/Humans.Budget.Contracts/Humans.Budget.Contracts.csproj
@@ -1,14 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <!--
-    Budget's cross-section surface, carved into its own leaf project because two of its
-    consumers are still in Base: TicketQueryService (Humans.Application) reads the active
-    year to compute the break-even summary, and TicketingBudgetSyncJob
-    (Humans.Infrastructure/Jobs) drives the nightly actuals sync — recurring jobs are named
-    by concrete type in Shell's UseHumansRecurringJobs roll-call and there is no discovery
-    seam for them yet (design §15.6b). A Contracts/ *folder* inside Humans.Budget would
+    Budget's cross-section surface, carved into its own leaf project because one of its
+    consumers is still in Base: TicketQueryService (Humans.Application) reads the active year
+    to compute the break-even summary. A Contracts/ *folder* inside Humans.Budget would
     therefore make Base reference a section and cycle (design §15.5b,
-    memory/architecture/section-project-cycle-fix.md).
+    memory/architecture/section-project-cycle-fix.md). The second Base consumer,
+    TicketingBudgetSyncJob, moved into Humans.Budget/Contracts/ at G5 lane 5b-3
+    (nobodies-collective/Humans#866), taking Humans.Infrastructure's reference to this leaf
+    with it.
 
     It holds the read surface the other sections actually call — four methods out of
     IBudgetService's forty — plus the DTOs and the two enums that appear in their
