+++ src/Sections/Humans.Budget.Contracts/ITicketingBudgetService.cs
@@ -5,9 +5,9 @@ namespace Humans.Budget.Contracts;
 /// <summary>
 /// The one thing the Tickets→Budget bridge exposes outside the section: materializing
 /// completed weeks of ticket sales into budget line items. Driven by
-/// <c>TicketingBudgetSyncJob</c>, which stays in <c>Humans.Infrastructure/Jobs</c>
-/// because recurring jobs are named by concrete type in Shell's roll-call
-/// (design §15.6b).
+/// <c>TicketingBudgetSyncJob</c>, which moved to <c>Humans.Budget/Contracts/</c> at G5 lane
+/// 5b-3 (nobodies-collective/Humans#866) — so this member is now section-internal in fact.
+/// Left public rather than shrunk: interface surface is Peter's call.
 /// </summary>
 /// <remarks>
 /// Projection refresh, the projection-parameter write and the virtual weekly forecast are
