+++ src/Sections/Humans.Budget/Contracts/TicketingBudgetSyncJob.cs
@@ -1,14 +1,21 @@
 using Hangfire;
 using Humans.Application.Interfaces;
-using Humans.Budget.Contracts;
 using Microsoft.Extensions.Logging;
 
-namespace Humans.Infrastructure.Jobs;
+namespace Humans.Budget.Contracts;
 
 /// <summary>
 /// Hangfire recurring job that materializes ticket sales actuals into budget line items.
 /// Runs daily at 04:30. Finds the active budget year's ticketing group and syncs completed weeks.
 /// </summary>
+/// <remarks>
+/// Moved out of <c>Humans.Infrastructure/Jobs</c> at G5 lane 5b-3
+/// (nobodies-collective/Humans#866). Budget, not Tickets: both collaborators
+/// (<see cref="ITicketingBudgetService"/>, <see cref="IBudgetServiceRead"/>) are Budget's,
+/// the rows it writes are budget line items, and the job id is <c>budget-ticketing-sync</c>.
+/// It sits under <c>Contracts/</c> because Shell names the concrete type at registration and
+/// HUM0034 makes every other public type in a section assembly an error.
+/// </remarks>
 [DisableConcurrentExecution(timeoutInSeconds: 300)]
 public class TicketingBudgetSyncJob(
     ITicketingBudgetService ticketingBudgetService,
