+++ src/Sections/Humans.Budget/Docs/Budget.md
@@ -220,7 +220,7 @@ Stored as string via `HasConversion<string>()`.
 ## Cross-Section Dependencies
 
 - **Teams:** `ITeamService.GetBudgetableTeamsAsync` / `ITeamService.GetEffectiveBudgetCoordinatorTeamIdsAsync` — narrow cross-section reads for team lookups and coordinator-scope resolution.
-- **Tickets:** none inbound. The Tickets→Budget bridge (`TicketingBudgetService`) is Budget's own service — it reads paid orders via `ITicketServiceRead` and calls the section's internal `IBudgetService` (`SyncTicketingActualsAsync` / `RefreshTicketingProjectionsAsync` / `UpdateTicketingProjectionAsync` / `GetTicketingProjectionEntriesAsync`). Budget has no code path that reads Tickets tables directly. Only `SyncActualsAsync` is public, on the contracts leaf, for the nightly `TicketingBudgetSyncJob`. (The dedicated `ITicketingBudgetRepository` added for PR #545b was removed in #815.)
+- **Tickets:** none inbound. The Tickets→Budget bridge (`TicketingBudgetService`) is Budget's own service — it reads paid orders via `ITicketServiceRead` and calls the section's internal `IBudgetService` (`SyncTicketingActualsAsync` / `RefreshTicketingProjectionsAsync` / `UpdateTicketingProjectionAsync` / `GetTicketingProjectionEntriesAsync`). Budget has no code path that reads Tickets tables directly. Only `SyncActualsAsync` is public, on the contracts leaf, for the nightly `TicketingBudgetSyncJob` (`Humans.Budget/Contracts/` since G5 lane 5b-3, so that member is now section-internal in fact). (The dedicated `ITicketingBudgetRepository` added for PR #545b was removed in #815.)
 - **Users/Identity:** `IUserServiceRead.GetUserInfosAsync` — actor display names for audit log. `IUserService.GetMergedSourceIdsAsync` — chain-follow merge tombstones on `BudgetAuditLog` GDPR export so source-attributed entries surface for the fold target.
 - **Admin:** Budget year lifecycle management is restricted to FinanceAdmin and Admin.
 
@@ -230,7 +230,7 @@ Stored as string via `HasConversion<string>()`.
 **Owned tables:** `budget_years`, `budget_groups`, `budget_categories`, `budget_line_items`, `budget_audit_logs`, `ticketing_projections`
 **Status:** (A) Migrated (peterdrier/Humans PR for issue nobodies-collective/Humans#544, 2026-04-22). Moved into its own project `src/Sections/Humans.Budget` at G5 (nobodies-collective/Humans#866).
 
-- Everything but `Section`, `BudgetResource` and the migrations is `internal` — HUM0034 enforces it. The cross-section surface is the leaf project `Humans.Budget.Contracts`: `IBudgetServiceRead` (four methods), `ITicketingBudgetService` (one), `IBudgetDemoSeeder` (one), the DTOs those name, and the `BudgetYearStatus` / `ExpenditureType` enums. The leaf is a project rather than a `Contracts/` folder because two consumers are in Base — `TicketQueryService` and `TicketingBudgetSyncJob`.
+- Everything but `Section`, `BudgetResource` and the migrations is `internal` — HUM0034 enforces it. The cross-section surface is the leaf project `Humans.Budget.Contracts`: `IBudgetServiceRead` (four methods), `ITicketingBudgetService` (one), `IBudgetDemoSeeder` (one), the DTOs those name, and the `BudgetYearStatus` / `ExpenditureType` enums. The leaf is a project rather than a `Contracts/` folder because one consumer is still in Base — `TicketQueryService`. `TicketingBudgetSyncJob` was the second until G5 lane 5b-3 moved it into this project's `Contracts/` folder.
 - `BudgetService` lives in `Humans.Budget.Services` and depends only on Application-layer abstractions. `IBudgetService` (internal) is the full surface on top of the read interface; it stays an interface because the ticketing bridge's unit tests substitute it.
 - `BudgetRepository` (impl `src/Sections/Humans.Budget/Data/BudgetRepository.cs`, §15b Singleton + `IDbContextFactory<BudgetDbContext>`) is the only file that touches budget tables via `DbContext`. `IBudgetRepository` exposes atomic per-method operations — multi-entity mutations (e.g. creating a year with its default groups / categories / projection row, or syncing ticketing actuals + re-materializing projected line items) are single repository methods that do all their work inside one short-lived `DbContext`.
 - **Decorator decision — no caching decorator.** Budget is admin-only, low-traffic. Same rationale as Governance / User / Feedback.
