+++ src/Sections/Humans.Gate/Contracts/GateRetentionJob.cs
@@ -1,11 +1,10 @@
 using Hangfire;
 using Humans.Application.Interfaces;
-using Humans.Gate.Contracts;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using NodaTime;
 
-namespace Humans.Infrastructure.Jobs;
+namespace Humans.Gate.Contracts;
 
 /// <summary>
 /// Retention purge for <c>gate_scan_events</c>. Gate scans are attendance /
@@ -13,6 +12,12 @@ namespace Humans.Infrastructure.Jobs;
 /// <c>Gate:RetentionDays</c> (default 365) are deleted daily. Set the value to
 /// 0 or below to disable the purge.
 /// </summary>
+/// <remarks>
+/// Moved out of <c>Humans.Infrastructure/Jobs</c> at G5 lane 5b-3
+/// (nobodies-collective/Humans#866). It sits under <c>Contracts/</c> because Shell names the
+/// concrete type at registration and HUM0034 makes every other public type in a section
+/// assembly an error.
+/// </remarks>
 [DisableConcurrentExecution(timeoutInSeconds: 300)]
 public sealed class GateRetentionJob(
     IGateScanRetention gateScans,
