+++ src/Sections/Humans.Gate/Controllers/GateController.cs
@@ -9,7 +9,7 @@ using Hangfire;
 using Humans.Application;
 using Humans.Application.Interfaces.Users;
 using Humans.Users.Contracts;
-using Humans.Infrastructure.Jobs;
+using Humans.Gate.Contracts;
 using Humans.UI.Authorization;
 using Humans.UI.Extensions;
 using Humans.UI.Models;
